+++ src/firebase.android.ts
@@ -299,9 +299,11 @@ firebase.init = arg => {
       arg = arg || {};
       initializeArguments = arg;
 
-      com.google.firebase.analytics.FirebaseAnalytics.getInstance(
-          appModule.android.context || com.tns.NativeScriptApplication.getInstance()
-      ).setAnalyticsCollectionEnabled(arg.analyticsCollectionEnabled !== false);
+      if (typeof (com.google.firebase.analytics) === "undefined") {
+        com.google.firebase.analytics.FirebaseAnalytics.getInstance(
+            appModule.android.context || com.tns.NativeScriptApplication.getInstance()
+        ).setAnalyticsCollectionEnabled(arg.analyticsCollectionEnabled !== false);
+      }
 
       // note that this only makes sense if crash reporting was disabled in AndroidManifest.xml
       if (arg.crashlyticsCollectionEnabled && typeof (com.crashlytics) !== "undefined" && typeof (com.crashlytics.android.Crashlytics) !== "undefined") {
@@ -309,10 +311,6 @@ firebase.init = arg => {
             appModule.android.context || com.tns.NativeScriptApplication.getInstance(),
             [new com.crashlytics.android.Crashlytics()]);
       }
-
-      com.google.firebase.analytics.FirebaseAnalytics.getInstance(
-          appModule.android.context || com.tns.NativeScriptApplication.getInstance()
-      ).setAnalyticsCollectionEnabled(arg.analyticsCollectionEnabled !== false);
 
       if (typeof (com.google.firebase.database) !== "undefined" && typeof (com.google.firebase.database.ServerValue) !== "undefined") {
         firebase.ServerValue = {